Remove function parameters
No need to have an easy built-in syntax for function literals with
parameters because there needs to be a helper for creating functions
that throw exceptions on parameter validation anyway, such as number of
parameters, type of parameters, preconditions, etc.

Deferred functions can be invoked, even though they might have symbols
in it, because they're evaluated at call time. It also allows for
optional dynamic binding, with the ability to inspect the function body.